Refactor code base, add documentation
