+++ server/bootstrap.peers
@@ -1,0 +1,2 @@
+/ip4/194.113.73.99/tcp/44981/p2p/QmUreqMsLeKVuoCJCQEGLwKZ8M5cogoPsQQcEYUj26kefi
+/ip4/209.151.148.27/tcp/44981/p2p/QmTnq5S4WZxoz9p1yP7JJBLXorgDYDqCnaKi97pTWUA3kQ
